requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4
aiosignal==1.2.0
alembic==1.8.1
anyio==3.7.1
async-timeout==3.0.1
asyncpg==0.26.0
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.5
cryptography==41.0.3
discord==1.7.3
discord-components-mirror==2.1.5
discord-voice==0.0.0
discord.py==2.2.2
exceptiongroup==1.2.0
frozenlist==1.3.0
greenlet==1.1.3
h11==0.14.0
httptools==0.6.1
idna==3.3
iniconfig==1.1.1
Mako==1.2.2
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.3
PyNaCl==1.5.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-asyncio==0.23.2
python-dotenv==0.20.0
PyYAML==6.0.1
requests==2.28.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
transliterate==1.10.2
typing_extensions==4.2.0
urllib3==1.26.9
watchfiles==0.21.0
websockets==3.4
yandex-music==2.2.0
yarl==1.7.2
